ISS-552 Supports workflow dispatch action event (#6)
The pixee/upload-tool-results-action has been updated to support the workflow_dispatch event in addition to the existing check_run and pull_request events.

It will solely handle the upload of the specified tool results file provided through the file input.

After reviewing the push event, it may not be necessary for SonarCloud, as SonarCloud is triggered when pushing and finishes between the check_run (already supported by the pixee action). However, it might be
an option to implement it when the analysis is different from SonarCloud, for example, Semgrep. But before running the pixee action, we would need something that creates the file and passes it in the `file` property.