✨ Add Support for Uploading Local Tool Results Instead of Fetching Results
When using `file` with any of the tools that support automatically
fetching results, the action would ignore the `file` input and instead
attempt to automatically fetch the results. The intention is for the
`file` input to replace the automatic fetching.

This causes some friction with Sonar, since we introduced two separate
result schemas for Sonar (issues and hotspots). When using `file`, the
user must specify whether the local file is a Sonar issues file or Sonar
security hotspots file.

/close #work